Working product

+++ drf_jwt_capstone_backend/replies/views.py
@@ -4,6 +4,8 @@ from .serializers import RepliesSerializer
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
+from rest_framework.permissions import IsAuthenticated
+from django.http import Http404
 
 # Create your views here.
 
@@ -11,15 +13,17 @@ from rest_framework import status
 class RepliesList(APIView):
 
     def get(self, request):
-        Replies = Replies.objects.all()
-        serializer = RepliesSerializer(Replies, many=True)
+        replies = Replies.objects.all()
+        serializer = RepliesSerializer(replies, many=True)
         return Response(serializer.data)
 
     def post(self, request):
-        serializer = RepliesSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        permission_classes = [IsAuthenticated]
+        if permission_classes == [IsAuthenticated]:
+            serializer = RepliesSerializer(data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -39,15 +43,33 @@ class RepliesDetail(APIView):
 
     # update
     def put(self, request, pk):
-        Replies = self.get_object(pk)
-        serializer = RepliesSerializer(Replies, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
+        permission_classes = [IsAuthenticated]
+        if permission_classes == [IsAuthenticated]:
+            Replies = self.get_object(pk)
+            serializer = RepliesSerializer(Replies, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     # delete
     def delete(self, request, pk):
-        Replies = self.get_object(pk)
-        Replies.delete()
+        permission_classes = [IsAuthenticated]
+        if permission_classes == [IsAuthenticated]:
+            Replies = self.get_object(pk)
+            Replies.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+class RepliesByComment(APIView):
+    
+    def get_objects_by_comment(self, comments_id):
+        try:
+            return Replies.objects.filter(comments_id = comments_id)
+        except Replies.DoesNotExist:
+            raise Http404
+
+
+    def get(self, request, comment_id):
+        replies = self.get_objects_by_comment(comment_id)
+        serializer = RepliesSerializer(replies, many=True)
+        return Response(serializer.data)
